Branch3-file_management

+++ controllers/stuff.js
@@ -1,28 +1,59 @@
+const fs = require('fs');
 const Thing = require('../models/Thing');
 
 // Pour la création d'un objet.
 exports.createThing = (req, res, next) => {
-    delete req.body._id;
+    const thingObject = JSON.parse(req.body.thing);
+    delete thingObject._id;
+    delete thingObject._userId;
     const thing = new Thing({
-        ...req.body // ... -> opérateur 'spread', il va aller copier les champs de req.body et les détailler un par un.
+        ...thingObject, // ... -> opérateur 'spread', il va aller copier les champs de req.body et les détailler un par un.
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     thing.save() // Enregistre l'objet dans la base.
     .then(() => res.status(201).json({message: 'Objet enregistré !'}))
-    .catch(error => res.status(400).json({error})); 
+    .catch(error => res.status(400).json({error}));
 };
 
 // Pour modifier un objet.
 exports.modifyThing = (req, res, next) => {
-    Thing.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet modifié !' }))
+    const thingObject = req.file ? {
+        ...JSON.parse(req.body.thing),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+    delete thingObject.userId;
+    Thing.findOne({_id: req.params.id})
+    .then((thing) => {
+        if(thing.userId != req.auth.userId){
+            res.status(401).json({ message: 'Non autorisé !' });
+        }
+        else{
+            Thing.updateOne({ _id: req.params.id }, {...thingObject, _id: req.params.id})
+            .then(() => res.status(200).json({ message: 'Objet modifié !' }))
+            .catch(() => res.status(401).json({ error }));
+        }
+    })
     .catch(error => res.status(400).json({ error }));
 };
 
 // Pour supprimer un objet.
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
-    .catch(error => res.status(400).json({ error }));
+    Thing.findOne({ _id: req.params.id })
+    .then(thing => {
+        if(thing.userId != req.auth.userId){ // On vérifie que la demande de suppression viens bien du créateur de l'objet.
+            res.status(401).json({ message: 'Non autorisé !' });
+        }
+        else{
+            const filename = thing.imageUrl.split('/images/')[1]; // On récup le nom du fichier.
+            fs.unlink(`images/${filename}`, () => { // 'unlink' pour supp le fichier sur le serveur de fichier.
+                Thing.deleteOne({_id: req.params.id}) // Delete l'objet dans la bdd.
+                .then(() => res.status(200).json({ message: 'Objet supprimé !' }))
+                .catch(error => res.status(400).json({ error }));
+            });
+        }
+    })
+    .catch(error => res.status(500).json({ error }));
 };
 
 // Pour récupérer un objet.
